Adds usage reporting to Wordpress example. (#7)
The metering agent is now deployed as a sidecar container in the
wordpress pod, and a Wordpress plugin handles reporting usage each time
a blog post is requested. The plugin also adds a dashboard widget that
reports metering agent status.

A new init container has been added that installs the Wordpress metering
plugin and copies the metering agent config to the proper location.

The metering agent is currently only configured to write to disk, not
Service Control. No secrets are required to run the app at this point.